Start moving autofix code into separate folder
